Change numbers to int except for divide, added comments

+++ calculator.py
@@ -26,9 +26,9 @@ def my_calculator():
         operator = tokens[0]
 
         if len(tokens) > 1:
-            num1 = float(tokens[1])
+            num1 = int(tokens[1])
             if len(tokens) > 2:
-                num2 = float(tokens[2])
+                num2 = int(tokens[2])
 
         # Here we start doing the math
         if operator == "+":
@@ -41,7 +41,7 @@ def my_calculator():
             print multiply(num1, num2)
 
         elif operator == "/":   # when adding error checking, check for division by 0
-            print divide(num1, num2)
+            print divide(float(num1), num2)
 
         elif operator == "square":  # consider adding 'or **'
             print square(num1)
@@ -57,6 +57,8 @@ def my_calculator():
 
         elif operator == 'q':
             calculator_on = False
+
+            #add else for invalid input
 
 
 my_calculator()
